# Meilensteine/Meilenstein_04/specs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Softwarepraktikum
  description: |-
    API spec for the server developed for the software technology internship at TU Chemnitz.
  license:
    name: AGPL V3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html#license-text
  version: "1.1.0"
servers:
  - url: swp.dczlabs.xyz:3130/api
  - url: localhost:3130
paths:
  /user:
    get:
      tags:
        - user
      responses:
        "200":
          description: user found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: user successfully created
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          $ref: "#/components/responses/ConflictError"
    put:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdatePartial"
      responses:
        "200":
          description: user got updated
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: user not found
  /user-edit/{userId}:
    parameters:
      - name: userId
        in: path
        description: ID of user to return
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdatePartial"
      responses:
        "200":
          description: user got deleted
        "404":
          description: user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
    delete:
      tags:
        - user
      responses:
        "200":
          description: user got deleted
        "404":
          description: user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
    get:
      tags:
        - user
      responses:
        "200":
          description: user found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /users:
    get:
      tags:
        - user
      responses:
        "200":
          description: all users returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /user/apitoken:
    get:
      tags:
        - user
      responses:
        "200":
          description: apitoken valid
        "401":
          description: apitoken not valid
    post:
      tags:
        - user
      security:
        - basicAuth: []
      responses:
        "200":
          description: login successfull. apiKey-cookie set.
        "401":
          description: login not successfull
    delete:
      tags:
        - user
      responses:
        "200":
          description: apitoken delted
        "401":
          description: apitoken not found or user not logged in
  /user/reset-password:
    get:
      tags:
        - user
      responses:
        "200":
          description: password reset request sent
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "403":
          $ref: "#/components/responses/ForbiddenError"
    post:
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
      responses:
        "200":
          description: password reset request created
  /user/reset-password/{username}:
    delete:
      tags:
        - user
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
            format: string
      responses:
        "200":
          description: password reset request deleted
        "404":
          description: password reset request not found
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    keyAuth:
      type: apiKey
      in: cookie
      name: apiKey
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - admin
        - email
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: Admin
        admin:
          type: boolean
        email:
          type: string
          example: admin@example.org
    UserCreate:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required:
            - passwd
          properties:
            passwd:
              type: string
    UserUpdatePartial:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        admin:
          type: boolean
        email:
          type: string
          example: admin@example.org
        passwd:
          type: string
...
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    ForbiddenError:
      description: User not allowed to execute this operation
    ConflictError:
      description: Send data conflicts with existing entry
security:
  - keyAuth: []